Python Project: Develop prediction models
Python Project: Develop prediction models by Kate Rogatina

In this project, we will develop several models that will predict the price of the car using the variables or features. This is just an estimate but should give us an objective idea of how much the car should cost.


Some questions we want to ask in this module


• Do I know if the dealer is oﬀering fair value for my trade-in?


• Do I know if I put a fair value on my car?


In data analytics, we often use Model Development to help us predict future observations from the data we have.


A model will help us understand the exact relationship between diﬀerent variables and how these variables are used to predict the result.