Analysis and prediction of market price of houses
In this Project I am in the role of a Data Analyst working at a real estate investment trust organization who wants to start investing in residential real estate. I am given a dataset containing detailed information about house prices in the region based on a number of property features, and it will be my job to analyze and predict the market price of houses given that information.