Visualization with Python Project by Kate
Visualization with Python by Kate (Introduction to Matplotlib and Line Plots)
Introduction
The aim of this Project is to introduce you to Matplotlib and creating Line Plots. 

The Dataset: Immigration to Canada from 1980 to 2013 

Dataset Source: International migration flows to and from selected countries - The 2015 revision. In this lab, we will focus on the Canadian immigration data.   We have already pre-processed the data, we will use the clean data saved in the csv format for this lab. The Canada Immigration dataset can be fetched from here.